# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AdminSys SaaS API",
    "version": "1.0.0",
    "description": "Standard SaaS Onboarding API for external integrations.",
    "contact": {
      "name": "API Support",
      "email": "support@adminsys.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.adminsys.com/api/v1",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-app-key",
        "description": "Unique identifier for the SaaS application (tenant)."
      },
      "Timestamp": {
        "type": "apiKey",
        "in": "header",
        "name": "x-timestamp",
        "description": "Current Unix timestamp in seconds. Used for signature verification."
      },
      "Nonce": {
        "type": "apiKey",
        "in": "header",
        "name": "x-nonce",
        "description": "A random string unique to this request to prevent replay attacks."
      },
      "Signature": {
        "type": "apiKey",
        "in": "header",
        "name": "x-signature",
        "description": "HMAC-SHA256 hash of (Method + Path + Timestamp + Nonce + Body). Signed with App Secret."
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": [],
      "Timestamp": [],
      "Nonce": [],
      "Signature": []
    }
  ],
  "paths": {
...
    "/api/admin/agents/reorder": {
      "post": {
        "tags": [
          "Admin - Agents"
        ],
        "summary": "Reorder AI agents",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/admin/agents/{id}": {
      "put": {
        "tags": [
          "Admin - Agents"
        ],
        "summary": "Update an AI agent",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated agent"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin - Agents"
        ],
        "summary": "Delete an AI agent",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/admin/agents/{id}/prompts": {
      "get": {
        "tags": [
          "Admin - Agents"
        ],
        "summary": "Get agent prompts",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of prompts"
          }
        }
      },
      "post": {
        "tags": [
          "Admin - Agents"
        ],
        "summary": "Upsert agent prompts",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated prompts"
          }
        }
      }
    },
...
    "/api/admin/ai-gateway/alerts/rule": {
      "get": {
        "tags": [
          "Admin - AI Gateway"
        ],
        "summary": "Get alert rule for app",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "appId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alert rule"
          }
        }
      }
    },
...
    "/api/admin/ai-gateway/alerts/preview": {
      "get": {
        "tags": [
          "Admin - AI Gateway"
        ],
        "summary": "Preview alert metrics",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "appId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Metrics preview"
          }
        }
      }
    },
...
    "/api/admin/ai-gateway/usage/today": {
      "get": {
        "tags": [
          "Admin - AI Gateway"
        ],
        "summary": "Get today's usage",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "appId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage data"
          }
        }
      }
    },
...
    "/api/admin/ai-gateway/policies": {
      "get": {
        "tags": [
          "Admin - AI Gateway"
        ],
        "summary": "List policies",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "appId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of policies"
          }
        }
      }
    },
...
  },
...
}